requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
configparser==5.2.0
docker-pycreds==0.4.0
gitdb==4.0.9
GitPython==3.1.26
idna==3.3
numpy==1.21.5
pandas==1.3.5
pathtools==0.1.2
Pillow==9.0.0
promise==2.3
protobuf==3.19.4
psutil==5.9.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
scipy==1.7.3
sentry-sdk==1.5.4
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
subprocess32==3.5.4
termcolor==1.1.0
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.8
wandb==0.12.9
yaspin==2.1.0
